Update makemenu to support arguments.
Use tuple to add arguments.
The first item within the tupel needs to be a function.
All other items are asumed to be the arguments to the funktion executed when selected.
If the function return the True value the menu will be closed and the previous menu will be shown.
Multiple actions can be bundled within a function

Also added a Filebrowser example to show the power of this new menubuilder.
You can enter any Directory and view the contents of files.